backend/requirements.txt
flask==3.0.3
gunicorn==22.0.0
google-cloud-storage==2.17.0
requests==2.32.3

